requirements.txt
# Python 第三方依赖包
# 注意：需要 Python 3.10+ 才能安装 mcp 包
httpx>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
mcp>=1.0.0
fastmcp>=0.1.0
psutil>=5.9.0